fix: optimizing database data type on migration files
- price column on products table changed from integer and decimal(10,2)/(12,2) to decimal(15,2)
- default varchar length on migration files changed from 255 to 191
- fixing merchants image storage path on MerchantController.php

+++ database/migrations/2025_11_27_050844_create_jasas_table.php
@@ -13,8 +13,8 @@ return new class extends Migration {
             $table->string('title');
             $table->string('slug', 150)->unique()->nullable();
             $table->text('description')->nullable();
-            $table->integer('fixed_price')->default(0);
-            $table->integer('base_price')->default(0);
+            $table->decimal('fixed_price', 15, 2)->default(0);
+            $table->decimal('base_price', 15, 2)->default(0);
             $table->enum('delivery_type', ['online', 'on-site', 'in-store'])->default('in-store');
             $table->enum('service_type_booking', ['keranjang', 'booking', 'konsultasi'])->default('keranjang');
             $table->string('location_address')->nullable();
